+++ motimo-api/src/main/java/kr/co/api/goal/docs/GoalControllerSwagger.java
@@ -7,9 +7,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.UUID;
 import kr.co.api.goal.rqrs.GoalCreateRq;
 import kr.co.api.goal.rqrs.GoalIdRs;
-import kr.co.api.goal.rqrs.GoalListRs;
 import kr.co.api.goal.rqrs.GoalUpdateRq;
-import kr.co.api.goal.rqrs.GoalWithSubGoalTodoRs;
 import org.springframework.web.bind.annotation.PathVariable;
 
 @Tag(name = "목표 API", description = "목표 관련 API 목록입니다")
@@ -29,10 +27,4 @@ public interface GoalControllerSwagger {
             @ApiResponse(responseCode = "404", description = "목표를 찾을 수 없음")
     })
     GoalIdRs goalComplete(UUID userId, @PathVariable UUID goalId);
-
-    @Operation(summary = "목표 목록 API", description = "목표 목록을 조회합니다.")
-    GoalListRs getGoalList(UUID userId);
-
-    @Operation(summary = "목표 투두 목록 API", description = "목표에 해당하는 세부 목표와 투두 목록을 조회합니다.")
-    GoalWithSubGoalTodoRs getTodoListByGoal(@PathVariable UUID goalId);
 }
